backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
chardet==3.0.4
charset-normalizer==3.4.7
click==8.3.2
deepl==1.29.0
exceptiongroup==1.3.1
fastapi==0.135.3
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.33.1
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.43.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
chardet==3.0.4
charset-normalizer==3.4.7
click==8.3.2
deepl==1.29.0
exceptiongroup==1.3.1
fastapi==0.135.3
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.33.1
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.43.0
